cp: multithreaded comparisons
Still pretty messy, and some considerable refactor / restructuring could
be used. But this appears to work well enough for a checkpoint. The idea
is this will replace jsoncmp.h / .cpp as more abstract and intentional
classes. For now leaving in the dead/redundant logic for
comparisons before finalizing a good structure.